requirements.txt
click
pycryptodome
pyotp
qrcode
requests
